backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Flask",
    "gunicorn",
    "google-genai",
    "python-dotenv",
    "google-cloud-pubsub",
    "flask-cors",
]
